# pos/openapi/RmsSupportHub.Pos.Agent.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RMS+ POS Agent API",
    "description": "Destination-owned foundation contract for the per-device Windows POS Agent. The Agent exposes a fixed HTTPS loopback endpoint for direct Support Hub browser access. Windows Negotiate supplies the authenticated identity, local Built-in Administrators membership authorizes protected reads and typed Windows-service and RMS database operations, and no Support Hub backend relay is involved. The Agent exposes read-only device/configuration/RMS visibility plus allow-listed service control and server-owned RMS database backup/restore workflows. Downloader, maintenance, and artifact capabilities remain typed, principal-scoped, and fail-closed. Mutation tokens are short-lived, one-use, and bound to the authenticated principal, exact Origin, registered operation, target, method, and path.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://rms-pos-agent.localhost:5001",
      "description": "Fixed HTTPS loopback Agent origin"
    }
  ],
  "paths": {
...
    "/api/v1/diagnostic-console/preview/{targetId}": {
      "get": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Preview a fixed Diagnostic Console run",
        "description": "Previews only a logical target from the fixed Agent manifest. The executable, arguments, working directory, child environment, timeout, and output bounds are resolved server-side; arbitrary process execution is not accepted.",
        "operationId": "PreviewDiagnosticConsoleRun",
        "parameters": [
          {
            "name": "targetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Agent returned the typed DiagnosticConsolePreviewDto preview.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticConsolePreviewDto"
                },
                "example": {
                  "previewId": "opaque-preview-id",
                  "target": "branchServerApi",
                  "ready": false,
                  "state": "notAttempted",
                  "blockers": [
                    "console_executable_unavailable"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The Agent rejected a non-canonical host with host_rejected or a non-HTTPS request with https_required; the response uses the safe Agent problem-details contract.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "The request host is not accepted.",
                  "status": 400,
                  "code": "host_rejected",
                  "correlationId": "example-correlation-id"
                }
              }
            }
          },
          "401": {
            "description": "The Windows authentication middleware issued a Negotiate challenge. This framework response is bodyless and is not guaranteed to contain Agent problem details.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Negotiate challenge emitted by the Windows authentication middleware.",
                "schema": {
                  "type": "string",
                  "example": "Negotiate"
                }
              }
            }
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostic-console/runs": {
      "post": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Start an authorized fixed Diagnostic Console run",
        "description": "Accepts only a bounded typed request after server-owned validation. Destructive or side-effecting dispatch requires the exact-origin, administrator, one-use method/path mutation-token, challenge/confirmation, and idempotency boundaries; ambiguous outcomes are retained and never retried automatically.",
        "operationId": "StartDiagnosticConsoleRun",
        "requestBody": {
          "description": "The browser supplies only logical IDs, exact confirmation where required, and a bounded idempotency key. Server-owned paths, credentials, SQL, and service targets are not accepted.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticConsoleStartRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The Agent accepted a fixed manifest diagnostic run with bounded, separately redacted output artifacts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticConsoleRunDto"
                }
              }
            }
          },
          "400": {
            "description": "The Agent rejected the typed request, challenge, configuration, or bounded idempotency key with safe problem details or typed NotAttempted state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "The Windows authentication middleware issued a Negotiate challenge for this protected mutation.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Negotiate challenge emitted by the Windows authentication middleware.",
                "schema": {
                  "type": "string",
                  "example": "Negotiate"
                }
              }
            }
          },
          "403": {
            "description": "Authorization, exact-origin transport, SID resolution, or the one-use mutation-token boundary rejected the mutation request.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          },
          "500": {
            "description": "The Agent returned a safe generic server-error response without exception, credential, path, or SQL details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostic-console/runs/{operationId}": {
      "get": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Read one principal-scoped Diagnostic Console run",
        "description": "Reads principal-scoped, sanitized progress and outcome for one retained Agent operation. Operation IDs and artifact capabilities are opaque; raw paths, credentials, SQL, service names, and exception text never cross the response boundary.",
        "operationId": "GetDiagnosticConsoleRun",
        "parameters": [
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Agent returned the retained typed DiagnosticConsoleRunDto state and opaque artifact capabilities.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticConsoleRunDto"
                },
                "example": {
                  "operationId": "opaque-operation-id",
                  "state": "running",
                  "progressPercent": 50
                }
              }
            }
          },
          "404": {
            "description": "The opaque operation identifier is not retained for the authenticated principal."
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentProblemDetailsDto": {
        "required": [
          "type",
          "title",
          "status"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Problem type identifier for the stable Agent error contract."
          },
          "title": {
            "type": "string",
            "description": "Safe human-readable explanation of why the Agent rejected the request."
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "HTTP status code returned by the Agent.",
            "format": "int32"
          },
          "code": {
            "type": [
              "null",
              "string"
            ],
            "description": "Optional stable machine-readable Agent problem code used by clients; it never carries raw identity or credential data."
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Optional request correlation identifier echoed by the Agent for diagnostics."
          }
        },
        "description": "Safe application/problem+json error contract. Code is a stable problem code and correlationId may identify the request; sensitive identity and machine details are not included."
      },
...
      "DiagnosticConsolePreviewDto": {
        "required": [
          "previewId",
          "target",
          "ready",
          "displayName",
          "state",
          "maxWallTimeSeconds",
          "maxOutputBytes",
          "maxOutputLines",
          "preconditions",
          "blockers",
          "confirmationPhrase",
          "expiresAtUtc"
        ],
        "type": "object",
        "properties": {
          "previewId": {
            "type": "string",
            "description": "Opaque Agent-issued identifier scoped to the authenticated principal and typed workflow."
          },
          "target": {
            "description": "Fixed server-owned diagnostic manifest target; no browser-selected executable or arguments are accepted.",
            "$ref": "#/components/schemas/DiagnosticConsoleTargetDto"
          },
          "ready": {
            "type": "boolean",
            "description": "Boolean outcome or capability flag produced by the server-owned boundary."
          },
          "displayName": {
            "type": "string",
            "description": "Server-owned logical label or profile value; it is not a browser-selected path or endpoint."
          },
          "state": {
            "description": "Typed diagnostic preview state.",
            "$ref": "#/components/schemas/DiagnosticConsoleRunStateDto"
          },
          "maxWallTimeSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed.",
            "format": "int32"
          },
          "maxOutputBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed.",
            "format": "int32"
          },
          "maxOutputLines": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed.",
            "format": "int32"
          },
          "preconditions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bounded server-owned logical evidence or precondition labels; raw machine targets are omitted."
          },
          "blockers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Bounded server-owned logical evidence or precondition labels; raw machine targets are omitted."
          },
          "confirmationPhrase": {
            "type": "string",
            "description": "Exact server-owned typed confirmation value required for this workflow boundary."
          },
          "expiresAtUtc": {
            "type": "string",
            "description": "UTC lifecycle timestamp produced and enforced by the Agent.",
            "format": "date-time"
          }
        },
        "description": "Preview of one fixed Agent diagnostic manifest target with bounded process and output policy."
      },
      "DiagnosticConsoleResultDto": {
        "required": [
          "stdoutArtifactId",
          "stderrArtifactId",
          "stdoutBytes",
          "stderrBytes",
          "stdoutLines",
          "stderrLines",
          "outputTruncated",
          "redactionApplied",
          "exitCode",
          "detail"
        ],
        "type": "object",
        "properties": {
          "stdoutArtifactId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "stderrArtifactId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "stdoutBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed.",
            "format": "int32"
          },
          "stderrBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed.",
            "format": "int32"
          },
          "stdoutLines": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed.",
            "format": "int32"
          },
          "stderrLines": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed.",
            "format": "int32"
          },
          "outputTruncated": {
            "type": "boolean",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "redactionApplied": {
            "type": "boolean",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "exitCode": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed.",
            "format": "int32"
          },
          "detail": {
            "type": "string",
            "description": "Safe operator detail without credentials, raw paths, SQL, shell text, private keys, or exception data."
          }
        },
        "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
      },
      "DiagnosticConsoleRunDto": {
        "required": [
          "operationId",
          "target",
          "state",
          "outcome",
          "progressPercent",
          "stage",
          "detail",
          "startedAtUtc",
          "completedAtUtc",
          "result",
          "errorCode",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "description": "Opaque Agent-issued identifier scoped to the authenticated principal and typed workflow."
          },
          "target": {
            "description": "Fixed server-owned diagnostic manifest target used by this operation.",
            "$ref": "#/components/schemas/DiagnosticConsoleTargetDto"
          },
          "state": {
            "description": "Typed diagnostic operation lifecycle state.",
            "$ref": "#/components/schemas/DiagnosticConsoleRunStateDto"
          },
          "outcome": {
            "description": "Typed diagnostic operation outcome truth.",
            "$ref": "#/components/schemas/DiagnosticConsoleRunStateDto"
          },
          "progressPercent": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Bounded server-reported progress percentage for the typed operation.",
            "format": "int32"
          },
          "stage": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "detail": {
            "type": "string",
            "description": "Safe operator detail without credentials, raw paths, SQL, shell text, private keys, or exception data."
          },
          "startedAtUtc": {
            "type": "string",
            "description": "UTC lifecycle timestamp produced and enforced by the Agent.",
            "format": "date-time"
          },
          "completedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "description": "UTC lifecycle timestamp produced and enforced by the Agent.",
            "format": "date-time"
          },
          "result": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DiagnosticConsoleResultDto"
              }
            ],
            "description": "Separate bounded stdout/stderr artifact metadata, when a result exists."
          },
          "errorCode": {
            "type": [
              "null",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "correlationId": {
            "type": "string",
            "description": "Safe request correlation identifier for operator diagnostics."
          }
        },
        "description": "Principal-scoped fixed diagnostic operation state with safe detail and opaque output artifacts."
      },
      "DiagnosticConsoleRunStateDto": {
        "enum": [
          "notAttempted",
          "accepted",
          "running",
          "succeeded",
          "failed",
          "timedOut",
          "cancelled",
          "partial",
          "outcomeUnknown"
        ],
        "description": "Typed server-owned state classification; unknown and recovery states are preserved explicitly."
      },
      "DiagnosticConsoleStartRequestDto": {
        "required": [
          "previewId",
          "typedConfirmation",
          "idempotencyKey"
        ],
        "type": "object",
        "properties": {
          "previewId": {
            "type": "string",
            "description": "Opaque Agent-issued identifier scoped to the authenticated principal and typed workflow."
          },
          "typedConfirmation": {
            "type": "string",
            "description": "Exact server-owned typed confirmation value required for this workflow boundary."
          },
          "idempotencyKey": {
            "type": "string",
            "description": "Bounded caller-generated key used to prevent duplicate execution of the same typed request."
          }
        },
        "description": "Typed request for one prior fixed diagnostic preview. It contains no process path, shell, executable, argument, or environment input."
      },
      "DiagnosticConsoleTargetDto": {
        "enum": [
          "branchServerApi",
          "cashierServerApi",
          "serviceManager",
          "cashierUi"
        ],
        "description": "Server-owned logical label or profile value; it is not a browser-selected path or endpoint."
      },
...
    },
    "securitySchemes": {
      "WindowsNegotiate": {
        "type": "http",
        "description": "Windows Integrated Authentication using Negotiate. The browser/OS supplies credentials with a credentialed cross-origin request; clients must rely on the browser's integrated credential flow.",
        "scheme": "negotiate"
      }
    }
  },
...
}